requirements.txt
beautifulsoup4==4.12.2
fastapi==0.100.0
nltk==3.8.1
numpy==1.25.0
openai==0.27.8
pandas==2.0.3
pydantic==1.10.10
pymongo==4.4.0
redis==4.6.0
Requests==2.31.0
spacy==3.5.4
SQLAlchemy==2.0.17
ta==0.10.2
thinc==8.1.10
tiktoken==0.4.0
websocket_client==0.53.0
yfinance==0.2.22
